Handle pluggable transports with care
We should return bridges supporting pluggable transports as specified
in the configuration file, to the best of our ability.

When we fulfill a request for bridges we should try to return at
least the minimum number of pluggable transports as defined in the
configuration file (and no more than the max, if defined). We
accomplish this by iterating over the list of bridges we retrieve
from the ring and reordering it to meet the min/max values. This
allows us to return regular bridges as well as obfsproxy bridges,
with a single request/response. bug #8614.