Throw an exception if we can't return the requested num of bridges
This is an exceptional case and will allow for us to more easily handle
this case in the future if we want to take special action as a result.

For example, if we are unable to provide the requested number of bridges
because too many are blocked in the specified country, then we can return
a string ID which maps to a translation that we can display to the user
explaining what happened.