Add notice and instructions about using the PTTBB
Users are adding pluggable transports to the TBB (which doesn't support them).
We add a notice on the front page stating the necessity of downloading the PTTBB
and we add instructions on the bridges page which describe how a user can
determine if they are using the TBB and failing. Bugfix on #9156.