Parse descriptors/documents/statuses more strictly
The spec (Tor directory protocol) describes how the various
files are formated. We should follow the guidelines more closely
because it will result in logs which show if a Bridge Authority
is providing partially corrupt data or data in an invalid form.
We presently rely heavily on the source Bridge Authority to
provide sane files.