Do not print "bridge" on the bridge line
Vidalia does not need/want/ask that the bridge line start with "bridge". Users expect
to be able to copy and paste what they see on their screen and that it will work.
Presently, this is not the case. #5851 states that we print "bridge" so that users
can easily copy and paste the line directly into their torrc, however most users do not
do this, so we should not add this string. TorButton will preserve this functionality.

This patch can be reverted if the functionality in TBB3.x is changed.